Use Inconsolata as a Web Font
Linux distributions typically don't have Inconsolata installed and fall
back to some default monospace font, which may not look good with the
rest of the fonts in the report. For example, this happens with Ubuntu's
Liberation Mono.

As we already have a downloadable Lato font in the report, adding
another isn't a big deal. The report looks good enough if downloading
the font fails.