Fix "missing dependency" errors when building within Gradle
KMP declares the same outputs twice, on the webpack task and on the
lifecycle task that depends on it. When running this build as included
into main Gradle build, something pulls both these tasks into the build
and sets off missing dependency alert. As a workaround, we're adding a
dummy dependency.